[API] Partial process instance start action request
Add support for identifying and buffering the 'instance start'
action requests, as well as replying the appropriate response.

A 'history' of actions is kept where actions are identified by
unique 'action_id' field. This is structured as a FIFO to keep
relative order of requested actions.
'action_id's can be reused as long as existing action with that
id has completed (timestamp is updated - None means not done -
Some(timestamp) is the completion timestamp). If an 'action_id'
can be reused, the updated action is moved to the front of the
history FIFO.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>