refactoring the kernel configuration
The kernel related configuration is now part of the KernelConfig
struct. This includes the kernel path, the kernel start address
and the kernel command line address.
Removed the kernel path and the kernel cmdline from the MachineCfg
since they are not needed.
Inside the VMM, the kernel config is optional as in time we will
add other boot sources. Setting the kernel config is mandatory
for other operations (like adding a drive) and should be the
first operation done after starting firecracker.

Signed-off-by: Andreea Florescu <fandree@amazon.com>