virtio/net: add option to specify guest MAC addr
Refactored code in devices/src/virtio/net.rs, and added logic to allow
a specific MAC address to be selected for a guest network device. This
required the addition of a minimal config space for net devices, together
with the VIRTIO_NET_F_MAC flag, as specified in the Virtio standard.

Signed-off-by: Alexandru Agache <aagch@amazon.com>