Add support for interrogating Guest Instance information through the API
This commit defines and implements support for HTTP GET on URL '/'. The
HTTP response body will have Guest Instance information; for now, just
the guest vm state.

The Guest VM state is now tracked and a validation is also added to the
'InstanceStart' command to disallow attempting a boot while a Guest VM
instance is already booting or running. This fixes issue #198.

The newly supported API request does not follow the 'sync' request model.
Instead of being forwarded to the VMM thread and the response being passed
back and returned to the client, the Guest Instance information is shared
between the threads and the API thread has direct read access to it.
For this API request, the VMM thread does no work.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>