vmm: make Vm a required field of VMM
This is needed because the VM fd is used for configuring the VM
before starting the instance.
As part of this commit, the memory became an Option field so
that we can get the fd and set the guest memory only when we know
the size.
Also, split the setup function in logical functions.

Signed-off-by: Andreea Florescu <fandree@amazon.com>