NetworkInterfaceBody: reworked de/serialization
The de/serialization of this structure previously relied on two
local helper functions to handle the guest_mac: Option<MacAddr>
field. However, this led to buggy behaviour, as the guest_mac = None
case was not handled properly. We eliminate the local functions, and
rely on the implementation of Serialize and Deserialize for MacAddr
instead. Also added one more unit test function.

Signed-off-by: Alexandru Agache <aagch@amazon.com>