removed num_cpus crate
This crate was only used to get the number of processors on host
as it was needed as an upper bound to the vcpu_count. The problem is
that num_cpus functions are based on _SC_NPROCESSORS_ONLN from
sysconfig and this is not implemented as expected. Instead of returning
the number of processors on the host, it returns the number of processors
that the running process can use. When running firecracker with kcov,
this is limited to 1 (because kcov is pinned to 1 cpu), which causes
the tests on put_virtual_machine_configuration to fail.

Removed the upper limit for the vCPU count for now, when we have a good
solution for enforcing an upper limit, we can revisit this decision.

Signed-off-by: Andreea Florescu <fandree@amazon.com>