tests: added integration test for CPU Topology
This is the first test from the integration tests that uses ssh
to login into the microvm. For this purpose, a new class was
added to handle the ssh connection and passing commands (SSHConnection).

Also, the microvms with the "capability:ssh" need to have an SSH
configuration file and an ssh key in the fsfiles S3 directory. The path
of these files are saved as fields in the MicrovmSlot and used to
connect to the microvm.

The first test added is testing a microvm with 2 vCPUs and
hyperthreading disabled. It uses lscpu to get the configuration
of the guest and matches it with the expected topology.

Signed-off-by: Andreea Florescu <fandree@amazon.com>