tests: added separated dirs for build and ...
make build directories persistent accross tests. Previously,
the build would be done in the root repository directory as it is
the default path of the cargo build command. Created separate
directories for each build type (unit tests, debug, release and
kcov). This proves to be very helpful because we have access to
the release binary without rebuilding it before starting the
functional tests.

The firecracker binaries are created in custom folders in the
root session directory. For the release binary to be persistent
accross different tests, the test session root path is removed
only after all tests are done. Also, since the builds are separate,
there is no need to run cargo clean, which also decreases the test
time.

Added a host_tools module that should contain functions used
across many tests. The first helper added is the cargo_build.

Signed-off-by: Andreea Florescu <fandree@amazon.com>