virtio-net: coalesce RX interrupts delivered to the guest
Instead of kicking the guest driver after delivering each RX packet,
deliver as many packets are already available on the tap and there
is room for in the virtqueue and only then kick the guest driver.

This results in slightly higher latency but much higher throughput
and even less IRQ-handling guest cpu-overhead.

NOTE: it would be worth an experiment for setting up a timer for
kicking the guest. But that would need serious fine tuning.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>