MMDS: get data by path
The method get_data_by_path is to be used when the guest sends
an HTTP request to 169.254.169.254/${PATH}. The path is split
by "/" and the tokens are searched in the data store. If the
path does not match an object in the data store, NotFound Error
is thrown. Also, if one object does not have the type Value::String
or Value::Object, UnsupportedValueType Error is thrown.

The method returns a Vec of Strings with variable length:
- 0 if trying to list a file as a directory
- 1 if trying to get the value of one key
- x, where x is the number of keys in a sub-dictionary specified
  by path.

For more details, check the public Instance meta-data Service
documentation:

https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-instance-metadata.html

Signed-off-by: Andreea Florescu <fandree@amazon.com>