virtio/net: mmds_ns is no longer in a Box
The mmds_ns field of struct NetEpollHandler is now an Option that
directly contains an MmdsNetworkStack object, instead of an Option which
holds a Box<MmdsNetworkStack>. The initial approach was chosen because
we were planning to have a significant amount of statically allocated
memory in the MmdsNetworkStack struct, but right now it's around ~120
bytes. We'll put it back in the Box if it grows past a certain size.

Signed-off-by: Alexandru Agache <aagch@amazon.com>