vmm-refactoring: moved API related functions...
at the end of the VMM implementation.

* renamed put_virtual_machine_configuration to set_vm_configuration
* moved the logic of updating the machine configuration to
set_vm_configuration in order to separate VMM logic from API logic.
* added UpdateNotAllowedPostBoot to DriveError.

The logic behind these changes is to have a VMM public interface
that is not tied to the API. The VMM should not be aware of the
caller of its functions.

Signed-off-by: Andreea Florescu <fandree@amazon.com>