data_model: moved machine_config to vmm/vmm_config
As part of this change, also removed the function from the x86 crate
that was setting a cpu template based on the machine configuration.
Now the template is set when configuring the vcpu directly. This
change was needed because otherwise we would end up with a cyclic
dependency between the vmm and the x86 crate.

Signed-off-by: Andreea Florescu <fandree@amazon.com>