integ tests: use external clone() tool
Python's ctypes libc interface appears to be broken, causing a
deadlock in the process of cloning into the jailer. Fixed this by adding
a clone/exec tool, written in C, to the testing system.
Warning: two new issues remain to be solved:
1. memory_cop now needs to handle the case where the jailer process
exists first, and
2. test_shutdown now needs to watch the jailer PID via some other method
than waitpid(), since the jailer is no longer a direct child.

Signed-off-by: Dan Horobeanu <dhr@amazon.com>