api_server: logging improvements
Added logs for requests on /mmds. For these requests we cannot log
the body because it contains sensitive information (e.g. credentials).

Refactored the existing logs on requests so that for GET requests we
do not add the body to the log entry.

Rephrased some of the existing log entries.

Signed-off-by: Andreea Florescu <fandree@amazon.com>